Add python evaluator6, 7, 8 (#14)
* Init evaluator5

* Five card hand evaluator src & test

* Add TODO

* Delete unnecessary tables

* Fix syspath

* Split test_table.py into multiple files

* Minor fix

* Add script for unittest and README

* Add script for unittest and README (2)

* Fix typo

* Remove the currently unavailable hashtable6 from python source file

* Minor fix

* Fix bug in suit_hash

* Init evaluator6

* Update hashtable for noflush6

* Init evaluator5

* Five card hand evaluator src & test

* Add TODO

* Delete unnecessary tables

* Fix syspath

* Split test_table.py into multiple files

* Minor fix

* Add script for unittest and README

* Add script for unittest and README (2)

* Fix typo

* Remove the currently unavailable hashtable6 from python source file

* Minor fix

* Fix bug in suit_hash

* Init evaluator6

* Update hashtable for noflush6

* Fix hand list to contain leading zeros

* Fix bug. Baseline table has to be NOFLUSH5

* Follow changes in test_tables/utils.py

* Init evaluator7

* Init evaluator8

+++ python/test_tables/test_hashtable5.py
@@ -29,7 +29,7 @@ class TestNoFlush5Table(unittest.TestCase):
         self.USED[i] = 0
 
   def mark_four_of_a_kind(self):
-    # Order 13C2 lexicograhically
+    # Order 13C2 lexicographically
     self.gen_quinary(2, 2)
     for base in self.QUINARIES:
       idx = 0
@@ -147,7 +147,7 @@ class TestNoFlush5Table(unittest.TestCase):
   def mark_flush(self):
     # Selecting 5 cards in 13: 13C5
     # Need to exclude straight: -10
-    self.CUR_RANK += 13 * 12 * 11 * 10 * 9 / (5 * 4 * 3 * 2) - 10
+    self.CUR_RANK += int(13 * 12 * 11 * 10 * 9 / (5 * 4 * 3 * 2)) - 10
 
   def setUp(self):
     if not self.UPDATED:
